@@ -3,9 +3,10 @@
 
 use Spatie\Onboard\OnboardingManager;
 use Spatie\Onboard\OnboardingSteps;
+use Spatie\Onboard\Tests\User;
 
 beforeEach(function () {
-    $this->user = $this->mock('User');
+    $this->user = new User();
 });
 
 test('steps can be defined and configured', function () {
@@ -19,14 +20,14 @@ test('steps can be defined and configured', function () {
             return true;
         });
 
-    $this->assertEquals(1, $onboardingSteps->steps(new stdClass())->count());
+    $this->assertEquals(1, $onboardingSteps->steps(new User())->count());
 
-    $step = $onboardingSteps->steps(new stdClass())->first();
+    $step = $onboardingSteps->steps(new User())->first();
 
-    $this->assertEquals('/some/url', $step->link);
-    $this->assertEquals('Test This!', $step->cta);
-    $this->assertEquals('Test Step', $step->title);
-    $this->assertEquals('attribute', $step->another);
+    expect($step->link)->toBe('/some/url')
+        ->and($step->cta)->toBe('Test This!')
+        ->and($step->title)->toBe('Test Step')
+        ->and($step->another)->toBe('attribute');
 });
 
 test('is in progress when all steps are incomplete', function () {
@@ -36,8 +37,8 @@ test('is in progress when all steps are incomplete', function () {
 
     $onboarding = new OnboardingManager($this->user, $onboardingSteps);
 
-    $this->assertTrue($onboarding->inProgress());
-    $this->assertFalse($onboarding->finished());
+    expect($onboarding->inProgress())->toBeTrue()
+        ->and($onboarding->finished())->toBeFalse();
 });
 
 test('is finished when all steps are complete', function () {
@@ -49,8 +50,8 @@ test('is finished when all steps are complete', function () {
 
     $onboarding = new OnboardingManager($this->user, $onboardingSteps);
 
-    $this->assertTrue($onboarding->finished());
-    $this->assertFalse($onboarding->inProgress());
+    expect($onboarding->finished())->toBeTrue()
+        ->and($onboarding->inProgress())->toBeFalse();
 });
 
 test('it returns the correct next unfinished step', function () {
@@ -77,9 +78,9 @@ test('it returns the correct next unfinished step', function () {
 
     $nextStep = $onboarding->nextUnfinishedStep();
 
-    $this->assertNotNull($nextStep);
-    $this->assertEquals("Step 2", $nextStep->title);
-    $this->assertEquals("/step-2", $nextStep->link);
+    expect($nextStep)->not()->toBeNull()
+        ->and($nextStep->title)->toBe("Step 2")
+        ->and($nextStep->link)->toBe("/step-2");
 });
 
 test('nextUnfinishedStep returns null if all steps are completed', function () {
@@ -103,24 +104,71 @@ test('nextUnfinishedStep returns null if all steps are completed', function () {
 
     $nextStep = $onboarding->nextUnfinishedStep();
 
-    $this->assertNull($nextStep);
+    expect($nextStep)->toBeNull();
 });
 
 test('the proper object gets passed into completion callback', function () {
-    $user = $this->mock('User');
+    $user = $this->mock(User::class);
     $user->shouldReceive('testMe')->once();
 
     $onboardingSteps = new OnboardingSteps();
     $onboardingSteps->addStep('Test Step')
-        ->completeIf(function ($user) {
+        ->completeIf(function ($model) {
             // if this gets called, it ensures that the right object was passed through.
-            $user->testMe();
+            $model->testMe();
 
             return true;
         });
 
     $onboarding = new OnboardingManager($user, $onboardingSteps);
 
-    // Calling finished() will triger the completeIf callback.
-    $this->assertTrue($onboarding->finished());
+    expect($onboarding->finished())->toBeTrue();
+});
+
+it('can get percentage completed', function () {
+    $onboardingSteps = new OnboardingSteps();
+    $onboardingSteps->addStep('Step 1')
+        ->link("/step-1")
+        ->completeIf(function () {
+            return true;
+        });
+
+    $onboardingSteps->addStep('Step 2')
+        ->link("/step-2")
+        ->completeIf(function () {
+            return false;
+        });
+
+    $onboarding = new OnboardingManager($this->user, $onboardingSteps);
+
+    expect($onboarding->percentageCompleted())->toBe(50.0);
+});
+
+it('will only run complete callbacks once', function () {
+    $called = 0;
+
+    $onboardingSteps = new OnboardingSteps();
+    $onboardingSteps->addStep('Step 1')
+        ->link("/step-1")
+        ->completeIf(function () use (&$called) {
+            $called++;
+            return true;
+        });
+
+    $onboardingSteps->addStep('Step 2')
+        ->link("/step-2")
+        ->completeIf(function () {
+            return false;
+        });
+
+    $onboarding = new OnboardingManager($this->user, $onboardingSteps);
+
+    $onboarding->finished();
+
+    expect($called)->toBe(1);
+
+    $onboarding->finished();
+
+    expect($called)->toBe(1)
+        ->and($onboarding->finished())->toBeFalse();
 });
